cloud_run/service1/app/requirements.txt
requests
ptvsd # Required for debugging.
pandas
pyarrow
fastparquet
google-cloud-storage
gcsfs
google-cloud-pubsub